Am having issues committing multiple files, so this is commit number 2 of 2
Most of the changes relate to an icon that can now be added to each row on the species grid to link to a page of the user's choosing. There are 4 news options on the edit tab for setting this up.
There is also a change which allows the system to create a sample for each occurrence on the shorewatch project. I have also checked a normal dynamic_sample_occurrence form and the code did not alter the behaviour of old code in my tests.